Add Parser.java, Storage.java, Ui.java, Tasklist.java to implement more OOP
